Rewrites the discovery mechanism to circumvent the non-blocking natur… (#6)
* Rewrites the discovery mechanism to circumvent the non-blocking nature of the UPNP SSDP library.
Unintentioned exits before responses from the UPNP SSDP search has been received will thus no longer happen.

This has been implemented through the addition of two classes Discoverer and Device. The Discoverer
class contains the main business logic to make the UPNP SSDP search blocking (by timeouts), while the
Device class is just a basic data storage for all interesting information for a device, which currently
only is the address and what status code it responded with.
The Discoverer class is implemented as an event dispatcher. For every response from the UPNP SSDP search it
will itself emit its own discovery event with the found device and a list of all devices. The duration of
the discovery is customizable. It defaults to 5000 ms, but can be set to anything. If it is set to a value
<= 0 the discovery will continue indefinitely. When the discovery phase ends it stops the UPNP SSDP listener

The heos module now exposes the Discoverer class, the discover function and a simpler findAny function. The
latter two functions can be used if the use case is sufficiently simple to not require the flexibility
offered by the Discoverer class.

The discover function has had its footprint modified and it now takes a callback as an argument. The
callback us used for the callback for the discovery event emitted by the Discoverer object during
its discovery.

The findAny function resolves its promise with the first device found from the discoverer.

* Prettified

* Hardcode interval time for infinite search for devices

* Remove unnecessary class

* Added error message when no devices found

* Return the actual promise that resolves / rejects

The promise did not do anything. At the same time the underlying promise was not reachable so when it rejected it could not be catched. This solves that

* Decouple the stop sequence in discoverer for stoping infinite searches

* Fix bugs

Both with ```this``` not working correctly and with size that should be length

* Add helper function for clarity

* Changed let to const and removed verbosity

* Moved creation of error into a function to reduce code duplication.