feat: Parse message of HEOS response as object (#27)
resolves #19

BREAKING CHANGE: The message field in HeosResponse and HeosEvent is now an object instead of a string. If the message sent from the HEOS device is key-value pair formatted, it will be parsed into an object. The message object is an empty object if there was no message field in the response. Otherwise it is an object with two fields. It always has the field unparsed which is the raw string recieved. If the message can be parsed there is a field parsed with the key-value pairs in an object.